still getting scrollbar measurements

+++ TextEditor/TextEditorMeasurements.cs
@@ -4,11 +4,11 @@ namespace TextEditor;
 public struct TextEditorMeasurements
 {
     public double CharacterWidth { get; set; }
-    public int LineHeight { get; set; }
-    public int EditorWidth { get; set; }
-    public int EditorHeight { get; set; }
-    public int EditorLeft { get; set; }
-    public int EditorTop { get; set; }
+    public double LineHeight { get; set; }
+    public double EditorWidth { get; set; }
+    public double EditorHeight { get; set; }
+    public double EditorLeft { get; set; }
+    public double EditorTop { get; set; }
     /// <summary>"literal" as opposed to the "scrollHeight", this is the amount of width the y-axis scrollbar measures</summary>
     public double ScrollbarLiteralWidth { get; set; }
     /// <summary>"literal" as opposed to the "scrollWidth", this is the amount of height the x-axis scrollbar measures</summary>
